+++ lib/src/utils/constants.dart
@@ -1,2 +1,8 @@
-const String charHyphen = '-';
-const String charUnderscore = '_';
+const charHyphen = '-';
+const charUnderscore = '_';
+
+const keyLanguage = 'language';
+const keyExtlang = 'extlang';
+const keyScript = 'script';
+const keyRegion = 'region';
+const keyVariant = 'variant';